pyproject.toml
[project]
name = "cogniwise"
version = "0.1.0"
description = "Course platform with account profile and yookassa payment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-behaviors>=0.5.1",
    "django-environ>=0.12.0",
    "django-minio-storage>=0.5.8",
    "django-redis>=6.0.0",
    "django-resized>=1.0.3",
    "pillow>=11.3.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary,pool]>=3.2.10",
    "uvicorn>=0.35.0",
    "yookassa>=3.7.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-debug-toolbar>=6.0.0",
    "django-debug-toolbar-force>=0.2",
    "flake8>=7.3.0",
    "flake8-bandit>=4.1.1",
    "flake8-django>=1.1.1",
    "flake8-isort>=6.1.2",
    "flake8-pyproject>=1.2.3",
    "flake8-walrus>=1.2.0",
    "isort>=6.0.1",
    "mixer>=7.2.2",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]

[tool.flake8]
extend-select = ["I",]
ignore = ["E501", "W503", "DJ01", "S308", "S703"]
exclude = [
    "test_*.py",
    "*test_*.py",
    "conftest.py",
    "*/tests/",
    "*/migrations/*",
    ".git/",
]

[tool.mypy]
ignore_missing_imports = true
exclude = [
    "tests/*",
    "migrations/*",
]

[tool.isort]
profile = "black"
line_length = 88
